+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 #### Changed
 - Styles are applied sorted by specificity #5
 
+#### Fixed
+- Fixed nested style references
+
 [Commits](https://github.com/yonaskolb/XcodeGen/compare/0.1.0...0.2.0)
 
 ## 0.1.0